Validation routine now works
validation routine now spits out the proper data consistantly. There is
either left over data or garbage in the Tx buffer that needs to be
cleared so the python script dosnt stop at variflying page 1 saying the
device is not responding. other than that the code uploaded still works
as intended.